+++ ethpm/backends/ipfs.py
@@ -7,6 +7,12 @@ from eth_utils import import_string, to_bytes
 import ipfsapi
 
 from ethpm import ASSETS_DIR
+from ethpm._utils.ipfs import (
+    dummy_ipfs_pin,
+    extract_ipfs_path_from_uri,
+    generate_file_hash,
+    is_ipfs_uri,
+)
 from ethpm.backends.base import BaseURIBackend
 from ethpm.constants import (
     DEFAULT_IPFS_BACKEND,
@@ -14,12 +20,6 @@ from ethpm.constants import (
     IPFS_GATEWAY_PREFIX,
 )
 from ethpm.exceptions import CannotHandleURI, ValidationError
-from ethpm.utils.ipfs import (
-    dummy_ipfs_pin,
-    extract_ipfs_path_from_uri,
-    generate_file_hash,
-    is_ipfs_uri,
-)
 
 
 class BaseIPFSBackend(BaseURIBackend):
