+++ ethpm/backends/registry.py
@@ -1,13 +1,13 @@
 import os
 
+from eth_typing import URI
 from web3 import Web3
 from web3.providers.auto import load_provider_from_uri
 
+from ethpm._utils.registry import fetch_standard_registry_abi
+from ethpm._utils.uri import parse_registry_uri
 from ethpm.backends.base import BaseURIBackend
 from ethpm.constants import INFURA_API_KEY
-from ethpm.typing import URI
-from ethpm.utils.registry import fetch_standard_registry_abi
-from ethpm.utils.uri import parse_registry_uri
 from ethpm.validation import is_valid_registry_uri
 
 # TODO: Update registry ABI once ERC is finalized.
