+++ ethpm/deployments.py
@@ -4,8 +4,8 @@ from eth_utils import to_canonical_address
 from web3 import Web3
 from web3.eth import Contract
 
+from ethpm._utils.contract import validate_contract_name
 from ethpm.exceptions import ValidationError
-from ethpm.utils.contract import validate_contract_name
 
 
 class Deployments:
