+++ ethpm/package.py
@@ -2,10 +2,32 @@ import json
 from pathlib import Path
 from typing import Any, Dict, Generator, Optional, Tuple, Union
 
+from eth_typing import URI, Address, ContractName
 from eth_utils import to_canonical_address, to_text, to_tuple
 from web3 import Web3
 from web3.eth import Contract
 
+from ethpm._utils.backend import resolve_uri_contents
+from ethpm._utils.cache import cached_property
+from ethpm._utils.contract import (
+    generate_contract_factory_kwargs,
+    validate_contract_name,
+    validate_minimal_contract_factory_data,
+    validate_w3_instance,
+)
+from ethpm._utils.deployments import (
+    get_linked_deployments,
+    normalize_linked_references,
+    validate_deployments_tx_receipt,
+    validate_linked_references,
+)
+from ethpm._utils.manifest_validation import (
+    check_for_deployments,
+    validate_build_dependencies_are_present,
+    validate_manifest_against_schema,
+    validate_manifest_deployments,
+    validate_raw_manifest_format,
+)
 from ethpm.contract import LinkableContract
 from ethpm.dependencies import Dependencies
 from ethpm.deployments import Deployments
@@ -14,28 +36,6 @@ from ethpm.exceptions import (
     FailureToFetchIPFSAssetsError,
     InsufficientAssetsError,
     PyEthPMError,
-)
-from ethpm.typing import URI, Address, ContractName
-from ethpm.utils.backend import resolve_uri_contents
-from ethpm.utils.cache import cached_property
-from ethpm.utils.contract import (
-    generate_contract_factory_kwargs,
-    validate_contract_name,
-    validate_minimal_contract_factory_data,
-    validate_w3_instance,
-)
-from ethpm.utils.deployments import (
-    get_linked_deployments,
-    normalize_linked_references,
-    validate_deployments_tx_receipt,
-    validate_linked_references,
-)
-from ethpm.utils.manifest_validation import (
-    check_for_deployments,
-    validate_build_dependencies_are_present,
-    validate_manifest_against_schema,
-    validate_manifest_deployments,
-    validate_raw_manifest_format,
 )
 from ethpm.validation import (
     validate_address,
