+++ ethpm/tools/builder.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import tempfile
 from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Tuple
 
+from eth_typing import URI, HexStr, Manifest
 from eth_utils import (
     add_0x_prefix,
     is_hex,
@@ -17,15 +18,14 @@ from eth_utils.toolz import assoc, assoc_in, concat, curry, pipe
 from web3 import Web3
 
 from ethpm import Package
-from ethpm.backends.ipfs import BaseIPFSBackend
-from ethpm.exceptions import ManifestBuildingError, ValidationError
-from ethpm.typing import URI, HexStr, Manifest
-from ethpm.utils.chains import is_BIP122_block_uri
-from ethpm.utils.manifest_validation import (
+from ethpm._utils.chains import is_BIP122_block_uri
+from ethpm._utils.manifest_validation import (
     format_manifest,
     validate_manifest_against_schema,
 )
-from ethpm.utils.mappings import deep_merge_dicts
+from ethpm._utils.mappings import deep_merge_dicts
+from ethpm.backends.ipfs import BaseIPFSBackend
+from ethpm.exceptions import ManifestBuildingError, ValidationError
 from ethpm.validation import validate_address
