--- ethpm/typing/__init__.py
@@ -1,1 +1,0 @@
-from .types import URI, Address, ContractName, HexStr, Manifest  # noqa: F401
