+++ tests/ethpm/test_manifest_assets_are_valid.py
@@ -3,8 +3,8 @@ import json
 import pytest
 
 from ethpm import ASSETS_DIR
+from ethpm._utils.manifest_validation import validate_manifest_against_schema
 from ethpm.exceptions import InsufficientAssetsError
-from ethpm.utils.manifest_validation import validate_manifest_against_schema
 
 SOURCES_GLOB = "**/*.json"
