+++ tests/ethpm/utils/test_contract_utils.py
@@ -1,12 +1,12 @@
 import pytest
 
-from ethpm.exceptions import InsufficientAssetsError, ValidationError
-from ethpm.utils.contract import (
+from ethpm._utils.contract import (
     generate_contract_factory_kwargs,
     validate_contract_name,
     validate_minimal_contract_factory_data,
     validate_w3_instance,
 )
+from ethpm.exceptions import InsufficientAssetsError, ValidationError
 
 
 @pytest.mark.parametrize(
