+++ tests/ethpm/utils/test_manifest_validation_utils.py
@@ -1,14 +1,14 @@
 import pytest
 
 from ethpm import V2_PACKAGES_DIR
-from ethpm.exceptions import ValidationError
-from ethpm.utils.manifest_validation import (
+from ethpm._utils.manifest_validation import (
     extract_contract_types_from_deployments,
     validate_manifest_against_schema,
     validate_manifest_deployments,
     validate_manifest_exists,
     validate_raw_manifest_format,
 )
+from ethpm.exceptions import ValidationError
 from ethpm.validation import validate_manifest_version
